Add loading spinner functionality with `NgxSpinner` integration
Introduced a `BusyService` to manage a loading spinner's visibility during HTTP requests. Added a `loadingInterceptor` to display the spinner for active requests and hide it upon completion. Updated `app.component` to include `NgxSpinnerComponent` in the template and configured spinner animations/styles in `angular.json`. Expanded `package.json` to include the `ngx-spinner` dependency.